rename cucumber authentication package
rename cucumber authentication package to match naming convention

fix #10435

delete this
