+++ src/main/java/tech/jhipster/lite/generator/server/springboot/cucumber_authentication/infrastructure/primary/CucumberAuthenticationModuleConfiguration.java
@@ -1,11 +1,11 @@
-package tech.jhipster.lite.generator.server.springboot.cucumberauthentication.infrastructure.primary;
+package tech.jhipster.lite.generator.server.springboot.cucumber_authentication.infrastructure.primary;
 
 import static tech.jhipster.lite.generator.slug.domain.JHLiteFeatureSlug.*;
 import static tech.jhipster.lite.generator.slug.domain.JHLiteModuleSlug.*;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import tech.jhipster.lite.generator.server.springboot.cucumberauthentication.application.CucumberAuthenticationApplicationService;
+import tech.jhipster.lite.generator.server.springboot.cucumber_authentication.application.CucumberAuthenticationApplicationService;
 import tech.jhipster.lite.module.domain.resource.JHipsterModuleOrganization;
 import tech.jhipster.lite.module.domain.resource.JHipsterModulePropertiesDefinition;
 import tech.jhipster.lite.module.domain.resource.JHipsterModuleResource;
