+++ src/main/java/tech/jhipster/lite/generator/server/springboot/cucumber_authentication/package-info.java
@@ -1,0 +1,2 @@
+@tech.jhipster.lite.BusinessContext
+package tech.jhipster.lite.generator.server.springboot.cucumber_authentication;
