+++ src/test/java/tech/jhipster/lite/generator/server/springboot/cucumberauthentication/domain/CucumberAuthenticationModuleFactoryTest.java
@@ -5,6 +5,7 @@ import static tech.jhipster.lite.module.infrastructure.secondary.JHipsterModules
 import org.junit.jupiter.api.Test;
 import tech.jhipster.lite.TestFileUtils;
 import tech.jhipster.lite.UnitTest;
+import tech.jhipster.lite.generator.server.springboot.cucumber_authentication.domain.CucumberAuthenticationModuleFactory;
 import tech.jhipster.lite.module.domain.JHipsterModule;
 import tech.jhipster.lite.module.domain.JHipsterModulesFixture;
 import tech.jhipster.lite.module.domain.properties.JHipsterModuleProperties;
